pyproject.toml
[project]
name = "social-badge-be"
version = "0.1.0"
description = "social badge backend api"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "uuid-utils>=0.14.1",
    "bcrypt>=5.0.0",
    "redis[hiredis]>=7.4.0",
    "resend>=2.30.0",
    "structlog>=25.5.0",
    "rich>=15.0.0",
    "asgi-correlation-id>=4.3.4",
    "slowapi>=0.1.9",
    "python-jose[cryptography]>=3.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "fakeredis[lua]>=2.35.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "T20", "S", "ASYNC"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
